Create wordle.py
This script is the original implementation for this site. As you can see, it leaves a lot to be desired. It prints out all permutations of the inputted letters, not taking into account any of the rules of Wordle. You can't specify what letters must go where or filter out any potential words that have a letter in the wrong position. 

I'm including it here for completeness (and so I can finally delete it from my desktop). Since creating this site, I have not had a need for this script.